+++ app/src/main/java/geekbarains/material/model/PODRetrofitImpl.kt
@@ -1,4 +1,4 @@
-package geekbarains.material.ui.picture
+package geekbarains.material.model
 
 import com.google.gson.GsonBuilder
 import okhttp3.Interceptor
