+++ app/src/main/java/geekbarains/material/model/PODServerResponseData.kt
@@ -1,4 +1,4 @@
-package geekbarains.material.ui.picture
+package geekbarains.material.model
 
 import com.google.gson.annotations.SerializedName
