+++ app/src/main/java/geekbarains/material/model/PictureOfTheDayAPI.kt
@@ -1,4 +1,4 @@
-package geekbarains.material.ui.picture
+package geekbarains.material.model
 
 import retrofit2.Call
 import retrofit2.http.GET
