+++ app/src/main/java/geekbarains/material/ui/PictureDescriptionFragment.kt
@@ -1,4 +1,4 @@
-package geekbarains.material.ui.picture
+package geekbarains.material.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,14 +7,13 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import geekbarains.material.R
 
-class BottomSheetFragment : Fragment(){
-
+class PictureDescriptionFragment : Fragment(){
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         //return super.onCreateView(inflater, container, savedInstanceState)
-        return inflater.inflate(R.layout.bottom_sheet_layout,container,false)
+        return inflater.inflate(R.layout.picture_of_the_day_description,container,false)
     }
 }
