+++ app/src/main/res/layout/picture_of_the_day_description.xml
